requirements.txt
streamlit
google-generativeai
fpdf2
pandas
openpyxl
python-docx
python-dotenv 
